Modified exceptions for Helpers. Moved clear list teardown to base fixture. Refactored assertion for filters.
